requirements.txt
fastapi
uvicorn
requests
chromadb
ollama
rank-bm25
python-docx
pydantic


